+++ PlantsVsZombies/GameObject/Bullet/Bullet.h
@@ -4,15 +4,15 @@
 class Bullet : public PictureBox
 {
 protected:
-	int m_attackPower;
-	int m_moveSpeed;
+	float m_attackPower;
+	float m_moveSpeed;
 
 	void Move();
 public:
 	Bullet(Point p_pos, Size p_size, const wstring& p_imagePath);
 	virtual ~Bullet();
-	void Init(int p_ap, int p_ms);
+	void Init(float p_ap, float p_ms);
 	void Update();
 
-	int GetAttackPower();
+	float GetAttackPower();
 };
